Create README.md
plim-plim
plim-plim is a platform independent python script to download videos from brazilian TV broadcaster globo.com.

It uses instructions from an input file to search for the videos, and supports download of protected content, provided the user is a subscriber of globo.com.

Videos can be downloaded both as playlists or as local flash video files. In both cases, the downloaded files are formatted and named in a way they can readily be imported into XBMC video library or nicely shown in its video windows.